requirements.txt
numpy
pymongo
gymnasium
opencv-python-headless
PyYAML
flask
flask_cors
fancy_gym

